Restructured the project to use multiple files instead of putting all code within the main.rs file
